Add hive on tez read table test
